#2225 :truck: move mbw composable files in ergebnismeldung/MBW
