+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/MBW/mbwErgebnisAndWahlvorschlagMapper.ts
