+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/MBW/mwbStapelBCUtils.ts
