+++ wls-gui-wahllokalsystem/tests/composables/ergebnismeldung/MBW/mwbStapelBCUtils.spec.ts
@@ -14,7 +14,7 @@ import {
   vi,
 } from "vitest";
 
-import { useMwbStapelBCUtils } from "@/composables/ergebnisermittlung/mwbStapelBCUtils.ts";
+import { useMwbStapelBCUtils } from "@/composables/ergebnismeldung/MBW/mwbStapelBCUtils.ts";
 import { StapelArtEnum } from "@/types/ergebnismeldung/StapelArtEnum.ts";
 
 const mockDefinitions = vi.hoisted(() => ({
